requirements.txt
sarvamai
numpy
sounddevice
soundfile
python-dotenv
google-genai
requests
beautifulsoup4